Move Currency class to dao.dto package.
