+++ src/main/java/ru/mystamps/web/dao/dto/Currency.java
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package ru.mystamps.web.service.dto;
+package ru.mystamps.web.dao.dto;
 
 public enum Currency {
 	EUR, USD, GBP, RUB
